requirements.txt
Flask==3.0.0
twilio==8.10.0
google-generativeai==0.3.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
python-docx==1.1.0

